Merged in r268, r273, r276 and r277 from /branches/aughey-devel/ccl. Cleaned up the swig.i file a bit, and renamed make.sh to build-ruby.sh. Original messages (with some editing) follow:
Code for binding to external scripting languages

Adding pack() methods to duplicate the operator overloaded methods; this was 
done to make the CCL friendlier to scripting languages.

Created a GetBufferString method to get the buffer as a std::string

Using the newly created GetBufferString method for getting at the internal 
buffer.


git-svn-id: http://cigi.svn.sourceforge.net/svnroot/cigi/trunk/ccl@357 7298b0d4-bb18-0410-8e19-c8aa18815fbc